create HubAsset base class
